Move polynomial and trig_poly from util to common.
