+++ drake/common/polynomial.h
