+++ drake/common/test/trig_poly_test.cc
@@ -1,11 +1,11 @@
-#include "drake/util/trig_poly.h"
+#include "drake/common/trig_poly.h"
 
 #include <sstream>
 #include <map>
 
 #include "gtest/gtest.h"
 
-#include "drake/util/polynomial.h"
+#include "drake/common/polynomial.h"
 
 using namespace Eigen;
 using namespace std;
