+++ drake/examples/Pendulum/runDynamics.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 
 #include "drake/common/drake_path.h"
-#include "drake/util/polynomial.h"
+#include "drake/common/polynomial.h"
 #include "drake/examples/Pendulum/Pendulum.h"
 #include "drake/systems/Simulation.h"
 #include "drake/systems/plants/BotVisualizer.h"
