+++ drake/examples/kuka_iiwa_arm/run_kuka_iiwa_arm_dynamics.cc
@@ -7,7 +7,7 @@
 #include "drake/systems/cascade_system.h"
 #include "drake/systems/plants/BotVisualizer.h"
 #include "drake/systems/plants/RigidBodySystem.h"
-#include "drake/util/polynomial.h"
+#include "drake/common/polynomial.h"
 #include "drake/util/drakeAppUtil.h"
 
 using drake::RigidBodySystem;
