+++ drake/solvers/optimization.h
@@ -16,7 +16,7 @@
 #include "drake/solvers/mathematical_program.h"
 #include "drake/solvers/decision_variable.h"
 #include "drake/solvers/solution_result.h"
-#include "drake/util/polynomial.h"
+#include "drake/common/polynomial.h"
 
 namespace drake {
 namespace solvers {
