+++ drake/solvers/test/system_identification_test.cc
@@ -6,8 +6,8 @@
 
 #include "gtest/gtest.h"
 
-#include "drake/util/polynomial.h"
-#include "drake/util/trig_poly.h"
+#include "drake/common/polynomial.h"
+#include "drake/common/trig_poly.h"
 
 namespace drake {
 namespace solvers {
