+++ drake/systems/framework/test/basic_vector_test.cc
@@ -6,7 +6,7 @@
 #include "gtest/gtest.h"
 
 #include "drake/common/functional_form.h"
-#include "drake/util/polynomial.h"
+#include "drake/common/polynomial.h"
 #include "drake/util/eigen_matrix_compare.h"
 
 using drake::util::CompareMatrices;
