+++ drake/systems/trajectories/PiecewisePolynomial.h
@@ -2,7 +2,7 @@
 
 #include <Eigen/Core>
 #include "drake/systems/trajectories/PiecewisePolynomialBase.h"
-#include "drake/util/polynomial.h"
+#include "drake/common/polynomial.h"
 #include <vector>
 #include <random>
 #include <limits>
