+++ drake/util/test/CMakeLists.txt
@@ -27,17 +27,9 @@ add_executable(testConvexHull testConvexHull.cpp)
 target_link_libraries(testConvexHull drakeConvexHull)
 add_test(NAME testConvexHull COMMAND testConvexHull)
 
-add_executable(testTrigPoly testTrigPoly.cpp)
-target_link_libraries(testTrigPoly drakePolynomial ${GTEST_BOTH_LIBRARIES})
-add_test(NAME testTrigPoly COMMAND testTrigPoly)
-
 add_executable(testDrakeGeometryUtil testDrakeGeometryUtil.cpp)
 target_link_libraries(testDrakeGeometryUtil drakeGeometryUtil ${GTEST_BOTH_LIBRARIES})
 add_test(NAME testDrakeGeometryUtil COMMAND testDrakeGeometryUtil)
-
-add_executable(polynomial_test polynomial_test.cc)
-target_link_libraries(polynomial_test drakePolynomial ${GTEST_BOTH_LIBRARIES})
-add_test(NAME polynomial_test COMMAND polynomial_test)
 
 add_executable(eigen_matrix_compare_test eigen_matrix_compare_test.cc)
 target_link_libraries(eigen_matrix_compare_test ${GTEST_BOTH_LIBRARIES})
